+++ packages/insured-bridge-relayer/src/RelayerConfig.ts
@@ -40,6 +40,10 @@ const bridgePoolDeployData = {
   "0x04Fa0d235C4abf4BcF4787aF4CF447DE572eF828": {
     timestamp: 1635964053,
   },
+  // BADGER:
+  "0x3472a5a71965499acd81997a54bba8d852c6e53d": {
+    timestamp: 1636750354,
+  },
 };
 
 // For similar reasons to why we store BridgePool deployment times, we store BridgeDepositBox times for each L2 network
@@ -60,6 +64,7 @@ export interface BotModes {
   relayerEnabled: boolean; // Submits slow and fast relays
   disputerEnabled: boolean; // Submits disputes on pending relays with invalid params
   finalizerEnabled: boolean; // Resolves expired relays
+  l2FinalizerEnabled: boolean; // Facilitates L2->L1 bridging over the canonical roll-up bridge.
 }
 export class RelayerConfig {
   readonly bridgeAdmin: string;
@@ -67,11 +72,13 @@ export class RelayerConfig {
   readonly pollingDelay: number;
   readonly errorRetries: number;
   readonly errorRetriesTimeout: number;
+
   readonly whitelistedRelayL1Tokens: string[] = [];
   readonly whitelistedChainIds: number[] = [];
   readonly rateModels: { [key: string]: RateModel } = {};
   readonly activatedChainIds: number[];
   readonly l2BlockLookback: number;
+  readonly crossDomainFinalizationThreshold: number;
   readonly botModes: BotModes;
   readonly l1DeployData: { [key: string]: { timestamp: number } };
   readonly l2DeployData: { [key: string]: { blockNumber: number } };
@@ -85,22 +92,31 @@ export class RelayerConfig {
       RATE_MODELS,
       CHAIN_IDS,
       L2_BLOCK_LOOKBACK,
+      CROSS_DOMAIN_FINALIZATION_THRESHOLD,
       RELAYER_ENABLED,
       FINALIZER_ENABLED,
       DISPUTER_ENABLED,
+      CROSS_DOMAIN_FINALIZER_ENABLED,
       WHITELISTED_CHAIN_IDS,
       L1_DEPLOY_DATA,
       L2_DEPLOY_DATA,
     } = env;
 
-    this.botModes = {
-      relayerEnabled: RELAYER_ENABLED ? Boolean(RELAYER_ENABLED) : false,
-      disputerEnabled: DISPUTER_ENABLED ? Boolean(DISPUTER_ENABLED) : false,
-      finalizerEnabled: FINALIZER_ENABLED ? Boolean(FINALIZER_ENABLED) : false,
-    };
-
     assert(BRIDGE_ADMIN_ADDRESS, "BRIDGE_ADMIN_ADDRESS required");
     this.bridgeAdmin = toChecksumAddress(BRIDGE_ADMIN_ADDRESS);
+
+    this.botModes = {
+      relayerEnabled: RELAYER_ENABLED === "true" ? true : false,
+      disputerEnabled: DISPUTER_ENABLED === "true" ? true : false,
+      finalizerEnabled: FINALIZER_ENABLED === "true" ? true : false,
+      l2FinalizerEnabled: CROSS_DOMAIN_FINALIZER_ENABLED === "true" ? true : false,
+    };
+
+    this.crossDomainFinalizationThreshold = CROSS_DOMAIN_FINALIZATION_THRESHOLD
+      ? Number(CROSS_DOMAIN_FINALIZATION_THRESHOLD)
+      : 5;
+
+    assert(this.crossDomainFinalizationThreshold < 100, "CROSS_DOMAIN_FINALIZATION_THRESHOLD must be < 100");
 
     // L2 start block must be explicitly set unlike L1 due to how L2 nodes work. For best practices, we also should
     // constrain L1 start blocks but this hasn't been an issue empirically. As a data point, Arbitrum Infura has a