+++ packages/fx-tunnel-relayer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.11](https://github.com/UMAprotocol/protocol/compare/@uma/fx-tunnel-relayer@1.0.10...@uma/fx-tunnel-relayer@1.0.11) (2021-11-11)
+
+**Note:** Version bump only for package @uma/fx-tunnel-relayer
+
 ## [1.0.10](https://github.com/UMAprotocol/protocol/compare/@uma/fx-tunnel-relayer@1.0.9...@uma/fx-tunnel-relayer@1.0.10) (2021-11-09)
 
 **Note:** Version bump only for package @uma/fx-tunnel-relayer