+++ packages/monitors/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/UMAprotocol/protocol/compare/@uma/monitors@2.8.0...@uma/monitors@2.9.0) (2021-11-11)
+
+### Features
+
+- **monitors:** Add SkinnyOptimisticOracle monitor ([#3587](https://github.com/UMAprotocol/protocol/issues/3587)) ([0dcf04c](https://github.com/UMAprotocol/protocol/commit/0dcf04cc4ff46438c7f8be62264ced5447143735))
+
 # [2.8.0](https://github.com/UMAprotocol/protocol/compare/@uma/monitors@2.7.0...@uma/monitors@2.8.0) (2021-11-09)
 
 ### Features