+++ packages/scripts/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/UMAprotocol/protocol/compare/@uma/scripts@1.4.3...@uma/scripts@1.4.4) (2021-11-11)
+
+**Note:** Version bump only for package @uma/scripts
+
 ## [1.4.3](https://github.com/UMAprotocol/protocol/compare/@uma/scripts@1.4.2...@uma/scripts@1.4.3) (2021-11-09)
 
 **Note:** Version bump only for package @uma/scripts